network/private-endpoint: allow alias as well
The input Resource ID may be a resource id or service alias.  This
validation is preventing the service alias from being accepted.  The
service alias has a standard suffix that can be used to validate the
input.

https://docs.microsoft.com/en-us/rest/api/virtualnetwork/privatelinkservices/get#privatelinkservice
https://docs.microsoft.com/en-us/rest/api/virtualnetwork/privateendpoints/createorupdate#privatelinkserviceconnection

It's not clear from the documentation, but I've verified that using the
Alias is functional this way.